requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
fonttools==4.42.1
fpdf2==2.7.5
h11==0.14.0
idna==3.4
load-dotenv==0.1.0
pika==1.3.2
py3dns==4.0.0
pydantic==2.3.0
pydantic-core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2
validate-email-address==1
